+++ settings.php
@@ -262,6 +262,10 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/scsspre';
         $title = get_string('rawscsspre', 'theme_boost', null, true);
         $description = get_string('rawscsspre_desc', 'theme_boost', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS,
+            true
+        );
         $default = '';
         $setting = new admin_setting_scsscode($name, $title, $description, $default, PARAM_RAW);
         $setting->set_updatedcallback('theme_reset_all_caches');
@@ -271,6 +275,10 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/scss';
         $title = get_string('rawscss', 'theme_boost', null, true);
         $description = get_string('rawscss_desc', 'theme_boost', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP,
+            true
+        );
         $default = '';
         $setting = new admin_setting_scsscode($name, $title, $description, $default, PARAM_RAW);
         $setting->set_updatedcallback('theme_reset_all_caches');
@@ -512,6 +520,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/logo';
         $title = get_string('logosetting', 'theme_boost_union', null, true);
         $description = get_string('logosetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configstoredfile(
             $name,
             $title,
@@ -527,6 +538,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/logocompact';
         $title = get_string('logocompactsetting', 'theme_boost_union', null, true);
         $description = get_string('logocompactsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configstoredfile(
             $name,
             $title,
@@ -556,6 +570,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/favicon';
         $title = get_string('faviconsetting', 'theme_boost_union', null, true);
         $description = get_string('faviconsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configstoredfile(
             $name,
             $title,
@@ -577,6 +594,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/backgroundimage';
         $title = get_string('backgroundimagesetting', 'theme_boost_union', null, true);
         $description = get_string('backgroundimagesetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $setting = new admin_setting_configstoredfile(
             $name,
             $title,
@@ -592,6 +612,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/backgroundimageposition';
         $title = get_string('backgroundimagepositionsetting', 'theme_boost_union', null, true);
         $description = get_string('backgroundimagepositionsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $backgroundimagepositionoptions = [
                 THEME_BOOST_UNION_SETTING_IMAGEPOSITION_CENTER_CENTER =>
                         get_string('imageposition_center_center', 'theme_boost_union'),
@@ -632,6 +655,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/brandcolor';
         $title = get_string('brandcolor', 'theme_boost_union', null, true);
         $description = get_string('brandcolor_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -640,6 +666,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/brandedgraytones';
         $title = get_string('brandedgraytones', 'theme_boost_union', null, true);
         $description = get_string('brandedgraytones_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configselect($name, $title, $description, THEME_BOOST_UNION_SETTING_SELECT_NO, $yesnooption);
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -654,6 +683,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/linkcolor';
         $title = get_string('linkcolorsetting', 'theme_boost_union', null, true);
         $description = get_string('linkcolorsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -662,6 +694,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/buttonbrandcolor';
         $title = get_string('buttonbrandcolorsetting', 'theme_boost_union', null, true);
         $description = get_string('buttonbrandcolorsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -676,6 +711,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/bootstrapcolorsuccess';
         $title = get_string('bootstrapcolorsuccesssetting', 'theme_boost_union', null, true);
         $description = get_string('bootstrapcolorsuccesssetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -684,6 +722,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/bootstrapcolorinfo';
         $title = get_string('bootstrapcolorinfosetting', 'theme_boost_union', null, true);
         $description = get_string('bootstrapcolorinfosetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -692,6 +733,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/bootstrapcolorwarning';
         $title = get_string('bootstrapcolorwarningsetting', 'theme_boost_union', null, true);
         $description = get_string('bootstrapcolorwarningsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -700,6 +744,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/bootstrapcolordanger';
         $title = get_string('bootstrapcolordangersetting', 'theme_boost_union', null, true);
         $description = get_string('bootstrapcolordangersetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -732,6 +779,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/navbarcolor';
         $title = get_string('navbarcolorsetting', 'theme_boost_union', null, true);
         $description = get_string('navbarcolorsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $navbarcoloroptions = [
                 THEME_BOOST_UNION_SETTING_NAVBARCOLOR_LIGHT =>
                         get_string('navbarcolorsetting_light', 'theme_boost_union'),
@@ -755,6 +805,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/navbartint';
         $title = get_string('navbartintsetting', 'theme_boost_union', null, true);
         $description = get_string('navbartintsetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
         $setting->set_updatedcallback('theme_reset_all_caches');
         $tab->add($setting);
@@ -790,6 +843,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
             $name = 'theme_boost_union/activityiconcolor' . $purpose;
             $title = get_string('activityiconcolor' . $purpose . 'setting', 'theme_boost_union', null, true);
             $description = get_string('activityiconcolor' . $purpose . 'setting_desc', 'theme_boost_union', null, true);
+            $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+                THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMS
+            );
             $setting = new admin_setting_configcolourpicker($name, $title, $description, '');
             $setting->set_updatedcallback('theme_reset_all_caches');
             $tab->add($setting);
@@ -999,6 +1055,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/loginbackgroundimage';
         $title = get_string('loginbackgroundimage', 'theme_boost_union', null, true);
         $description = get_string('loginbackgroundimage_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_MWP
+        );
         $setting = new admin_setting_configstoredfile(
             $name,
             $title,
@@ -3957,6 +4016,9 @@ if ($hassiteconfig || has_capability('theme/boost_union:configure', context_syst
         $name = 'theme_boost_union/footnote';
         $title = get_string('footnotesetting', 'theme_boost_union', null, true);
         $description = get_string('footnotesetting_desc', 'theme_boost_union', null, true);
+        $description .= '<br />' . theme_boost_union_render_setting_override_notification(
+            THEME_BOOST_UNION_SETTING_TARGETPLATFORM_LMSMWP
+        );
         $setting = new admin_setting_confightmleditor($name, $title, $description, '');
         $tab->add($setting);